feat(graph-contracts): declare a package without the runtime kernel
Extracts the graph declaration surface from @voyant-travel/core into
@voyant-travel/graph-contracts: defineAdapter, defineExtension, defineModule,
defineProvider, defineProject, the port helpers, and the graph types.

An external adapter previously needed core — the DI container, registry, event
bus, saga, locking — in order to describe itself, and where core was declared a
peer dependency, so did everyone installing that adapter. The new package has
no dependencies.

core/project and core/project-facets re-export it, so none of the 239 files
importing them change.

definePlugin and VOYANT_GRAPH_PLUGIN_SCHEMA_VERSION are deprecated rather than
removed: RFC #3395 retired the classification, but external adapters still call
them and need a migration window to defineAdapter.

Refs #4075
