+++ packages/cruises/tsconfig.build.json
@@ -1159,6 +1159,8 @@
       ],
       "@voyant-travel/framework/self-host-export": ["../framework/dist/self-host-export.d.ts"],
       "@voyant-travel/framework/worker-job-host": ["../framework/dist/worker-job-host.d.ts"],
+      "@voyant-travel/graph-contracts": ["../graph-contracts/dist/index.d.ts"],
+      "@voyant-travel/graph-contracts/facets": ["../graph-contracts/dist/facets.d.ts"],
       "@voyant-travel/hono": ["../hono/dist/index.d.ts"],
       "@voyant-travel/hono/app": ["../hono/dist/app.d.ts"],
       "@voyant-travel/hono/auth": ["../hono/dist/auth/index.d.ts"],
