+++ scripts/checks/manifests/public-surface.json
@@ -11,6 +11,7 @@
     "@voyant-travel/admin-extension-sdk",
     "@voyant-travel/app-manifest",
     "@voyant-travel/custom-fields-contracts",
+    "@voyant-travel/graph-contracts",
     "@voyant-travel/ui",
     "@voyant-travel/webhook-delivery-contracts"
   ],
